+++ rss-puzzle/src/components/main/translationModule.ts
@@ -2,15 +2,15 @@ import { IWordData } from './interfaces';
 import { currentRound, currentSentenceIndex } from './main';
 
 export function displayTranslation(wordData: IWordData) {
- const button = document.getElementById(
+  const button = document.getElementById(
     'toggle-translation-button',
- ) as HTMLButtonElement;
+  ) as HTMLButtonElement;
 
- const translationSpan = document.createElement('span');
- translationSpan.classList.add('translation-hint');
- translationSpan.style.display = 'none';
+  const translationSpan = document.createElement('span');
+  translationSpan.classList.add('translation-hint');
+  translationSpan.style.display = 'none';
 
- if (translationSpan) {
+  if (translationSpan) {
     if (button.textContent === 'Скрыть перевод') {
       translationSpan.style.display = 'block';
     } else {
@@ -21,19 +21,19 @@ export function displayTranslation(wordData: IWordData) {
       button.textContent === 'Показать перевод'
         ? 'Показать перевод'
         : 'Скрыть перевод';
- }
+  }
 
- const translation =
+  const translation =
     wordData.rounds[currentRound - 1]?.words[currentSentenceIndex]
       ?.textExampleTranslate || '';
 
- translationSpan.textContent = translation;
+  translationSpan.textContent = translation;
 
- const translationContainer = document.getElementById('translation');
- if (translationContainer) {
+  const translationContainer = document.getElementById('translation');
+  if (translationContainer) {
     translationContainer.innerHTML = '';
     translationContainer.appendChild(translationSpan);
- } else {
+  } else {
     console.error('Element with ID "translation" not found');
- }
+  }
 }
